+++ frontend/connection/src/WebsocketConnection.tsx
@@ -29,12 +29,9 @@ import {
   BaseUriParts,
   buildHttpUri,
   buildWsUri,
-} from "@streamlit/utils"
-import {
-  IHostConfigResponse,
-  PerformanceEvents,
   Resolver,
-} from "@streamlit/lib"
+} from "@streamlit/utils"
+import { IHostConfigResponse } from "./types/Config"
 import { ForwardMsgCache } from "@streamlit/connection/src/ForwardMessageCache"
 import { SessionInfo } from "@streamlit/connection/src/SessionInfo"
 import { StreamlitEndpoints } from "@streamlit/connection/src/StreamlitEndpoints"
@@ -556,24 +553,24 @@ export class WebsocketConnection {
     const messageIndex = this.nextMessageIndex
     this.nextMessageIndex += 1
 
-    PerformanceEvents.record({ name: "BeginHandleMessage", messageIndex })
+    // PerformanceEvents.record({ name: "BeginHandleMessage", messageIndex })
 
     const encodedMsg = new Uint8Array(data)
     const msg = ForwardMsg.decode(encodedMsg)
 
-    PerformanceEvents.record({
-      name: "DecodedMessage",
-      messageIndex,
-      messageType: msg.type,
-      len: data.byteLength,
-    })
+    // PerformanceEvents.record({
+    //   name: "DecodedMessage",
+    //   messageIndex,
+    //   messageType: msg.type,
+    //   len: data.byteLength,
+    // })
 
     this.messageQueue[messageIndex] = await this.cache.processMessagePayload(
       msg,
       encodedMsg
     )
 
-    PerformanceEvents.record({ name: "GotCachedPayload", messageIndex })
+    // PerformanceEvents.record({ name: "GotCachedPayload", messageIndex })
 
     // Dispatch any pending messages in the queue. This may *not* result
     // in our just-decoded message being dispatched: if there are other
@@ -582,11 +579,11 @@ export class WebsocketConnection {
     while (this.lastDispatchedMessageIndex + 1 in this.messageQueue) {
       const dispatchMessageIndex = this.lastDispatchedMessageIndex + 1
       this.args.onMessage(this.messageQueue[dispatchMessageIndex])
-      PerformanceEvents.record({
-        name: "DispatchedMessage",
-        messageIndex: dispatchMessageIndex,
-        messageType: this.messageQueue[dispatchMessageIndex].type,
-      })
+      // PerformanceEvents.record({
+      //   name: "DispatchedMessage",
+      //   messageIndex: dispatchMessageIndex,
+      //   messageType: this.messageQueue[dispatchMessageIndex].type,
+      // })
       delete this.messageQueue[dispatchMessageIndex]
       this.lastDispatchedMessageIndex = dispatchMessageIndex
     }
