+++ frontend/lib/src/FileUploadClient.ts
@@ -20,11 +20,9 @@ import { v4 as uuidv4 } from "uuid"
 
 import { IFileURLs, IFileURLsResponse } from "@streamlit/proto"
 
-import { SessionInfo } from "@streamlit/utils/src/SessionInfo"
-import { StreamlitEndpoints } from "@streamlit/connection"
-import { logWarning } from "@streamlit/utils"
-import Resolver from "./util/Resolver"
-import { isValidFormId } from "../../utils/src/utils"
+import { SessionInfo, StreamlitEndpoints } from "@streamlit/connection"
+import { logWarning, Resolver } from "@streamlit/utils"
+import { isValidFormId } from "./util/utils"
 
 /** Common widget protobuf fields that are used by the FileUploadClient. */
 interface WidgetInfo {
