+++ frontend/lib/src/components/core/LibContext.tsx
@@ -17,26 +17,7 @@
 import React from "react"
 
 import { baseTheme, ThemeConfig } from "@streamlit/lib/src/theme"
-
-/**
- * The lib config contains various configurations that the host platform can
- * use to configure streamlit-lib frontend behavior. This should to be treated as part of the public
- * API, and changes need to be backwards-compatible meaning that an old host configuration
- * should still work with a new frontend versions.
- */
-export type LibConfig = {
-  /**
-   * the mapbox token that can be configured by a platform
-   */
-  mapboxToken?: string
-
-  /**
-   * Whether to disable the full screen mode all elements / widgets.
-   */
-  disableFullscreenMode?: boolean
-
-  enforceDownloadInNewTab?: boolean
-}
+import { LibConfig } from "@streamlit/connection"
 
 export interface LibContextProps {
   /** True if the app is in full-screen mode. */
