+++ frontend/lib/tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": "dist",
     "rootDir": "src",
   },
-  "include": ["src"],
+  "include": ["src", "../utils/src/Resolver.ts"],
   "exclude": ["node_modules/**/*"],
   "references": [
     { "path": "../connection" },
