+++ frontend/proto/tsconfig.json
@@ -5,6 +5,6 @@
     "rootDir": "src",
   },
   "include": ["src"],
-  "exclude": ["node_modules/**/*"],
+  "exclude": ["node_modules/**/*", "*/proto.js", "*/proto.d.ts"],
   "references": [],
 }
