+++ frontend/utils/package.json
@@ -19,11 +19,13 @@
     "lint": "eslint --ext .js --ext .jsx --ext .ts --ext .tsx --max-warnings 0 src"
   },
   "dependencies": {
+    "xxhashjs": "^0.2.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.21.0",
     "@babel/core": "^7.21.4",
     "@types/jest": "^27.4.3",
+    "@types/xxhashjs": "^0.2.2",
     "babel-jest": "^27.4.3",
     "babel-plugin-module-resolver": "^5.0.0",
     "eslint-plugin-streamlit-custom": "file:../eslint-plugin-streamlit-custom",
